Create PrimeCachePerformanceTest: * Extract targetedAccessTest from BytePrimeCacheTest
Update BytePrimeCacheTest:
* Refactor prime number list to be public and static, accessible from PerformanceTest class
